+++ src/config_manager.rs
@@ -36,6 +36,8 @@ pub struct Config {
 
     pub notify_of_unzipped_mods: bool,
     
+    pub game_download_dir: String,
+    
     #[serde(default)]
     pub pkg: Vec<Package>,
    
@@ -90,6 +92,7 @@ impl Default for Config {
             backup_mods_dir: backup_mods_dir.to_string_lossy().to_string(),
             show_execution_time: true,
             notify_of_unzipped_mods: false,
+            game_download_dir: dirs::download_dir().unwrap().to_str().unwrap().to_string(),
             sync_latest_game_version_file_every: 24,
             sync_mod_search_file_every: 24,
             pkg: Vec::default(),